+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.4.0 / 2019-12-24
+==================
+* [MISC] Updated tests to clear database file before each one.
+* [MISC] Added more test cases.
+* [MISC] Moved all tests to tests folder.
+* [HOTFIX] Changed the client to use window.location.host instead of localhost.
+* [HOTFIX] Made secure flag change connection string to wss.
+
 0.3.0 / 2019-12-24
 ==================
 * [FEATURE] Changed class properties to be strict and added the necessary setters/getters to be more strict.